# be/docs/swagger.yaml
definitions:
...
  codon-backend_internal_models.Chapter:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      order_index:
        type: integer
      subject_id:
        type: string
      updated_at:
        type: string
    type: object
  codon-backend_internal_models.ContentItem:
    properties:
      chapter:
        $ref: '#/definitions/codon-backend_internal_models.Chapter'
      chapter_id:
        type: string
      content_type:
        $ref: '#/definitions/codon-backend_internal_models.ContentType'
      course:
        $ref: '#/definitions/codon-backend_internal_models.Course'
      course_id:
        type: string
      created_at:
        type: string
      file_key:
        type: string
      hls_playlist_url:
        type: string
      id:
        type: string
      rating_avg:
        type: number
      rating_count:
        type: integer
      rejection_reason:
        type: string
      requires_subscription:
        type: boolean
      reviewed_at:
        type: string
      reviewed_by:
        type: string
      status:
        $ref: '#/definitions/codon-backend_internal_models.ContentStatus'
      title:
        type: string
      updated_at:
        type: string
      uploaded_by:
        type: string
      uploader:
        $ref: '#/definitions/codon-backend_internal_models.User'
      video_status:
        $ref: '#/definitions/codon-backend_internal_models.VideoStatus'
    type: object
...
  codon-backend_internal_models.ContentType:
    enum:
    - video
    - document
    type: string
    x-enum-varnames:
    - ContentVideo
    - ContentDocument
...
  internal_handlers.adminContentDetailResponse:
    properties:
      content:
        $ref: '#/definitions/codon-backend_internal_models.ContentItem'
      url:
        type: string
    type: object
...
  internal_handlers.createContentRequest:
    properties:
      chapter_id:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      content_type:
        enum:
        - video
        - document
        example: video
        type: string
      course_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      file_key:
        example: video/user-uuid/file-uuid.mp4
        type: string
      requires_subscription:
        example: true
        type: boolean
      title:
        example: Cell Biology — Lecture 1
        type: string
    type: object
...
  internal_handlers.errorResponse:
    properties:
      code:
        example: pool_too_small
        type: string
      details: {}
      error:
        example: descriptive error message
        type: string
    type: object
...
  internal_handlers.getContentResponse:
    properties:
      content:
        $ref: '#/definitions/codon-backend_internal_models.ContentItem'
      my_rating:
        type: integer
      url:
        type: string
    type: object
...
  internal_handlers.listContentResponse:
    properties:
      content:
        items:
          $ref: '#/definitions/codon-backend_internal_models.ContentItem'
        type: array
      my_ratings:
        additionalProperties:
          type: integer
        type: object
    type: object
...
  internal_handlers.messageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
...
  internal_handlers.rejectContentRequest:
    properties:
      reason:
        example: Questions contain formatting errors
        type: string
    type: object
...
  internal_handlers.updateContentRequest:
    properties:
      chapter_id:
        type: string
      file_key:
        type: string
      requires_subscription:
        type: boolean
      title:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/admin/content:
    get:
      description: Returns content items filterable by status. Defaults to pending_review.
      parameters:
      - default: pending_review
        description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by course ID
        in: query
        name: course_id
        type: string
      - description: Filter by chapter ID
        in: query
        name: chapter_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.listContentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: List content items by status (Admin)
      tags:
      - Admin
  /api/v1/admin/content/{id}:
    get:
      description: Returns full content item details regardless of status, including
        the file URL for review.
      parameters:
      - description: Content item UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.adminContentDetailResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get a content item (Admin)
      tags:
      - Admin
  /api/v1/admin/content/{id}/approve:
    post:
      description: Approves a pending_review content item. The teacher can then publish
        it.
      parameters:
      - description: Content item UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.messageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Approve a content item (Admin)
      tags:
      - Admin
  /api/v1/admin/content/{id}/reject:
    post:
      consumes:
      - application/json
      description: Rejects a pending_review content item with a mandatory reason.
      parameters:
      - description: Content item UUID
        in: path
        name: id
        required: true
        type: string
      - description: Rejection reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.rejectContentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Reject a content item (Admin)
      tags:
      - Admin
...
  /api/v1/chapters/{chapter_id}/content:
    get:
      description: Returns all published content items (videos, documents) for a specific
        chapter. Items the student hasn't unlocked (subscription/KYC required) are
        still listed but with file_key and hls_playlist_url stripped.
      parameters:
      - description: Chapter UUID
        in: path
        name: chapter_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.listContentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: List published content items for a chapter (Student)
      tags:
      - Content
  /api/v1/content/{id}:
    get:
      description: Returns the details of a published content item plus a playable
        URL. Requires an active subscription (and approved KYC, if the platform requires
        it) when the item has requires_subscription set.
      parameters:
      - description: Content item UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.getContentResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get a single published content item (Student)
      tags:
      - Content
...
  /api/v1/teacher/content:
    get:
      description: Returns all content items uploaded by the authenticated teacher.
        Teachers with can_manage_all_content=true see all content platform-wide.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.listContentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: List teacher's content items
      tags:
      - Teacher
    post:
      consumes:
      - application/json
      description: Creates a draft content item (video or document). For videos, a
        background transcoding job is automatically queued to convert the upload to
        HLS format. The file must be uploaded to S3 first via /uploads/presign.
      parameters:
      - description: Content details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.createContentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/codon-backend_internal_models.ContentItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Upload content item (Teacher)
      tags:
      - Teacher
  /api/v1/teacher/content/{id}:
    get:
      description: Returns full content item details. A teacher can view their own
        content (or all content if they have platform-wide permission).
      parameters:
      - description: Content item UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.adminContentDetailResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get a content item (Teacher)
      tags:
      - Teacher
    patch:
      consumes:
      - application/json
      description: Partially updates a content item's metadata.
      parameters:
      - description: Content item UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.updateContentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codon-backend_internal_models.ContentItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Update a content item (Teacher)
      tags:
      - Teacher
  /api/v1/teacher/content/{id}/publish:
    post:
      description: Transitions an approved content item to published, making it visible
        to students.
      parameters:
      - description: Content item UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.messageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Publish an approved content item (Teacher)
      tags:
      - Teacher
  /api/v1/teacher/content/{id}/submit-for-review:
    post:
      description: Transitions a draft or rejected content item to pending_review
        status.
      parameters:
      - description: Content item UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.messageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Submit content for admin review (Teacher)
      tags:
      - Teacher
...
swagger: "2.0"